Moved telemetry page into app directory
