+++ app/_redux/store.ts
@@ -1,11 +1,10 @@
 import { Action, applyMiddleware, configureStore, ThunkAction } from "@reduxjs/toolkit";
-import thunk from "redux-thunk";
-import classificationSlice from "../features/events/classificationSlice";
-import dataFilterSlice from "../features/events/dataFilterSlice";
-import eventQueryReducer from "../features/events/eventQuerySlice";
-import eventTelemetryReducer from "../features/events/eventTelemetrySlice";
-import lapTelemetryQuerySlice from "../features/events/lapTelemetryQuerySlice";
-import lapTelemetrySlice from "../features/events/lapTelemetrySlice";
+import classificationSlice from "./features/events/classificationSlice";
+import dataFilterSlice from "./features/events/dataFilterSlice";
+import eventQueryReducer from "./features/events/eventQuerySlice";
+import eventTelemetryReducer from "./features/events/eventTelemetrySlice";
+import lapTelemetryQuerySlice from "./features/events/lapTelemetryQuerySlice";
+import lapTelemetrySlice from "./features/events/lapTelemetrySlice";
 
 export const makeStore = () =>
   configureStore({
