+++ app/components/EventQuery/Combobox.tsx
@@ -1,3 +1,5 @@
+"use client";
+
 import { Fragment, useState } from "react";
 import { Combobox, Transition } from "@headlessui/react";
 import { CheckIcon, SelectorIcon } from "@heroicons/react/solid";
