+++ app/components/LapTelemetryQuery/Listbox.tsx
@@ -1,3 +1,5 @@
+"use client";
+
 import { Fragment } from "react";
 import { Listbox, Transition } from "@headlessui/react";
 import { CheckIcon, SelectorIcon } from "@heroicons/react/solid";
