+++ src/pages/_app.tsx
@@ -1,7 +1,7 @@
 import "../styles/globals.css";
 import type { AppProps } from "next/app";
 import { Provider } from "react-redux";
-import store from "../app/store";
+import store from "../../app/_redux/store";
 
 function MyApp({ Component, pageProps }: AppProps) {
   return (
